Add support for printing shiny pokemon
Shiny pokemon can either be printed randomly with a probability of 1/128,
or with the -s flag together with the -n command. Documentation for this
is also included in the README and man page.

Also fixed some shellcheck (https://www.shellcheck.net/) warnings and
ran shfmt (https://github.com/patrickvane/shfmt) on the script.

Fixes #5